requirements.txt
# CodeFinder Dependencies

# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
numpy>=1.21.0

# Vector Database
chromadb>=0.4.0

# Web Framework
flask>=2.3.0

# Configuration
PyYAML>=6.0.0

# Code Parsing
tree-sitter>=0.20.0
tree-sitter-cpp>=0.20.0

# Utilities
tqdm>=4.64.0
pathlib2>=2.3.0

# Optional: For fuzzy matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0

# Optional: For better text processing
regex>=2022.0.0